fix(firmware): PRO-13 native host build — use uint32_t for millis()-style OTA resolve timing
pio test -e native failed to compile: otaResolveTimedOut()'s static_assert
for millis() rollover (0xFFFFFFFF -> 5) relies on the subtraction wrapping
in a 32-bit-wide unsigned type. unsigned long is 32-bit on the ESP32
Arduino target (matching millis()'s real width) but 64-bit on the
[env:native] host test target, so on host the same literals just produce
a large positive number instead of wrapping, and the static_assert fails.

Switch otaResolveTimedOut()'s parameters, WebUIPlugin's stored
otaResolveStartMs field, and kOtaResolveTimeoutMs to explicit uint32_t so
the 32-bit wraparound semantics are identical on native and ESP32 — mirrors
OtaUpdateCheckPolicy.h's otaBackoffInterval, which already does this for
the same reason. millis() call sites are wrapped with static_cast<uint32_t>
where they feed the now-uint32_t field/parameter (value-preserving on the
real 32-bit-millis() target). Adds the (previously untracked) native test
file for this policy.

Fixes PRO-13

+++ src/display/plugins/OtaAsyncResolvePolicy.h
@@ -55,7 +55,14 @@ constexpr OtaResolveState otaResolveStateForDecision(OtaFlashDecision decision) 
 // this firmware does (see BeanManager), because unsigned wraparound makes
 // `nowMs - startMs` compute the correct elapsed duration even when nowMs has
 // wrapped past startMs.
-constexpr bool otaResolveTimedOut(unsigned long startMs, unsigned long nowMs, unsigned long timeoutMs) {
+// PRO-13 fix: these are millis()-flavored uint32_t, not `unsigned long`.
+// `unsigned long` is 32-bit on the ESP32 Arduino target (matching millis()'s
+// real 32-bit rollover) but 64-bit on the [env:native] host test target, so
+// the rollover static_assert below (0xFFFFFFFF wrapping to 5) only holds if
+// the subtraction is explicitly 32-bit-wide on BOTH platforms. Mirrors
+// OtaUpdateCheckPolicy.h's otaBackoffInterval, which already uses uint32_t
+// for the same millis()-math reason.
+constexpr bool otaResolveTimedOut(uint32_t startMs, uint32_t nowMs, uint32_t timeoutMs) {
     return (nowMs - startMs) >= timeoutMs;
 }
